Use traits to implement read_exact without code duplication
